requirements.txt
customtkinter
openai
SpeechRecognition
pygame
python-dotenv
psutil
pyaudio
pillow
pyinstaller
requests
numpy
